Finish useScrollCapture
