+++ docs/package.json
@@ -28,7 +28,7 @@
     "css-loader": "^0.28.7",
     "dotenv": "^7.0.0",
     "extract-react-types-loader": "^0.3.0",
-    "flow-bin": "^0.91.0",
+    "flow-bin": "^0.93.0",
     "html-webpack-plugin": "^3.2.0",
     "moment": "^2.20.1",
     "pretty-proptypes": "^0.5.0",
