+++ package.json
@@ -58,7 +58,7 @@
     "eslint-plugin-react": "^7.3.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "extract-react-types-loader": "^0.3.0",
-    "flow-bin": "^0.91.0",
+    "flow-bin": "^0.93.0",
     "gh-pages": "^1.1.0",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^25.1.0",
